dcos-oauth: fix tests
The dcos-oauth cmd expects user detail to be sent as json.
This request body cannot be left empty. This commit modifies the TestUsers
integration test to send minimal json along in the PUT request.

The cmd also marshals users to json through the User struct of which the fields are tagged
with 'json:omitempty'. This means empty or default values aren't in the response
JSON. I suspect the test data was recorded before the omitempty tag was added to
the User struct definition.

This commit updates the expected responses in the failing tests to match current behaviour.