dcos-oauth: fix email verification check
The verified check used to fail only if the input was a bool
that was also false. This fixes the condition so the check
also fails if the response is not a bool at all.

The external component's API would have to change to make this check fail
but since we're checking, we might as well guard against that.